refactor: Refactor import paths and file names in app and theme directories
- Update import path for `backstageTheme` to a folder in `App.tsx`
- Rename `HelloWorld.ts` to `index.ts` in the `theme` directory
